Refactor Morph Target system: update keyframe time type to double, enhance error handling, and improve serialization for deterministic output

+++ OloEngine/src/OloEngine/Animation/AnimationClip.h
@@ -41,7 +41,7 @@ namespace OloEngine
     // Morph target animation keyframe
     struct MorphTargetKeyframe
     {
-        f32 Time;
+        f64 Time;
         std::string TargetName;
         f32 Weight;
     };
